requirements.txt
flask==1.1.2
flask_restplus==0.13.0
python-dotenv==0.13.0
werkzeug==0.16.1
pyjwt==1.7.1
flask_sqlalchemy==2.4.1
sendgrid==6.2.2
PyMySQL==0.9.3
sqlacodegen==1.1.6
